Create PyGeoT_Part1
The PyGeoT_Part1 script reads the input file that contains the chemical composition of the waters that will be used (the user needs to name this file "InputWaters_database.txt".). At this point, the script also reads the file named "Groups", which is the file that contains the mineral assemblages that the user wants to try. PyGeoT writes an input file whit this information to execute GeoT/iGeoT. After, the script opens the output files and extracts all the information concerning classical geothermometers, statistical parameters, and temperatures computed by GeoT/iGeoT, and writes a file1.txt file. Finally, PyGeoT extracts the optimized concentrations of Al and Mg and creates a file2.txt file.