Create PyGeoT_Part2
The objectives of the PyGeoT_Part2 script are to merge both files of results (file1.txt and file2.txt) and also filter the results, by deleting the values of Al and Mg concentrations that are equal to the lower or upper limit of the specified value ranges. 
For this step, it is required that the user deletes manually the first line of the file1.txt and file2.txt, which is the titles of the columns.