Create PyGeoT_Part3
This is the last part of the script. The PyGeoT_Part3 results evaluate each output file and remove those for which the temperature estimation has a T-spread higher than ±20°C and an RMED/nmin value higher than 0.01. Then, PyGeoT selects, orders, and extracts 20 results with the lowest T-spread and RMED/nmin values. These are the results of GeoT/iGeoT for the best-clustering different mineral assemblages. PyGeoT then reads the associated 20 estimated temperatures and computes the corresponding median, standard deviation, and temperature spread. The median temperature is reported as the temperature estimated by PyGeoT (TPyGeoT). Finally, PyGeoT provides a graph synthesizing the results obtained with the 20 best-clustering mineral assemblages. 
The 20 temperature results are displayed in the upper part of the graph (y1 axis), with an error bar corresponding to the T-spread value computed by iGeoT. The values of optimized parameters are also included (y2 axis). The minerals involved in every assemblage with their respective identification number are shown in the lower part of the graph. The PyGeoT temperature estimate (median of the 20 temperature values) is shown on the right side of the graph, together with the associated statistical parameters (T-spread and standard deviation), temperature estimates from classical geothermometers as well as the reservoir temperature (e.g., bottom-hole temperature) if available.