Create Combinations_Generator
This script is used to generate mineral combinations (mineral assemblages). The user must consider that a minerals.txt file is necessary to execute this script. The file must contain the name of the mineral as is written in the thermodynamic database, all the minerals must be in one column. In the minerals.txt file, the user can modify the large list of minerals that we already provide. The idea behind the list that we use is to provide general guidance regarding the most common alteration minerals, especially when such knowledge is lacking for a particular geothermal site.